Fix all linter errors as per reviewers comment
