+++ README.md
@@ -11,7 +11,7 @@ If you installed git you can clone the code to your machine, or download a ZIP o
 
 Download the ZIP from this location, or run the following git command to clone the files to your machine:
 
-* Once the files are on your machine, open the html-css-javascript-getting-started folder in Visual Studio Code.
+* Once the files are on your machine, open the BookList-Microverse-Project folder in Visual Studio Code.
 
 * With the files open in Visual Studio Code, press the Go Live button at the bottom of the window to launch the files with Live Server.
