+++ awesome-books-with-ES6/modules/bookElement.js
@@ -1,0 +1,3 @@
+const bookContainer = document.querySelector('.book-container');
+
+export default bookContainer;
