+++ awesome-books-with-ES6/modules/time.js
@@ -1,0 +1,3 @@
+const time = document.querySelector('.time');
+
+export default time;
